refactor(example): proper package name for relay example app
