+++ examples/nostr-relay-example-application/src/main/java/org/tbk/nostr/example/relay/db/FlywayConfig.java
@@ -1,0 +1,20 @@
+package org.tbk.nostr.example.relay.db;
+
+import org.flywaydb.core.api.migration.JavaMigration;
+import org.springframework.boot.autoconfigure.flyway.FlywayConfigurationCustomizer;
+import org.springframework.context.ApplicationContext;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration(proxyBeanMethods = false)
+class FlywayConfig {
+    @Bean
+    FlywayConfigurationCustomizer flywayConfigurationCustomizer(ApplicationContext applicationContext) {
+        return configuration -> {
+            JavaMigration[] javaMigrations = applicationContext.getBeansOfType(JavaMigration.class)
+                    .values().toArray(JavaMigration[]::new);
+
+            configuration.javaMigrations(javaMigrations);
+        };
+    }
+}
