+++ examples/nostr-relay-example-application/src/main/java/org/tbk/nostr/example/relay/domain/event/EventEntityEvents.java
@@ -1,0 +1,13 @@
+package org.tbk.nostr.example.relay.domain.event;
+
+public final class EventEntityEvents {
+    private EventEntityEvents() {
+        throw new UnsupportedOperationException();
+    }
+
+    public record CreatedEvent(EventEntity.EventEntityId eventId) {
+    }
+
+    public record MarkDeletedEvent(EventEntity.EventEntityId eventId) {
+    }
+}
