--- examples/nostr-relay-example-application/src/main/java/org/tbk/nostr/relay/example/NostrRelayExampleSecurityConfig.java
@@ -1,55 +1,0 @@
-package org.tbk.nostr.relay.example;
-
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnWebApplication;
-import org.springframework.boot.autoconfigure.security.servlet.PathRequest;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.builders.WebSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configuration.WebSecurityCustomizer;
-import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
-import org.springframework.security.config.http.SessionCreationPolicy;
-import org.springframework.security.web.SecurityFilterChain;
-import org.springframework.security.web.firewall.StrictHttpFirewall;
-import org.springframework.security.web.header.writers.XXssProtectionHeaderWriter;
-
-import static org.springframework.security.web.util.matcher.AntPathRequestMatcher.antMatcher;
-
-@Slf4j
-@EnableWebSecurity
-@ConditionalOnWebApplication
-@Configuration(proxyBeanMethods = false)
-@RequiredArgsConstructor
-class NostrRelayExampleSecurityConfig implements WebSecurityCustomizer {
-
-    @Override
-    public void customize(WebSecurity web) {
-        web.httpFirewall(new StrictHttpFirewall());
-    }
-
-    @Bean
-    SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
-        http
-                .cors(AbstractHttpConfigurer::disable)
-                .csrf(AbstractHttpConfigurer::disable)
-                .sessionManagement(session -> session
-                        .sessionCreationPolicy(SessionCreationPolicy.NEVER)
-                        .sessionFixation().migrateSession()
-                )
-                .headers(headers -> headers
-                        .xssProtection(xss -> xss.headerValue(XXssProtectionHeaderWriter.HeaderValue.ENABLED_MODE_BLOCK))
-                )
-                .authorizeHttpRequests(authorize -> authorize
-                        .requestMatchers(PathRequest.toStaticResources().atCommonLocations()).permitAll()
-                        .requestMatchers(
-                                antMatcher("/")
-                        ).permitAll()
-                        .anyRequest().authenticated()
-                );
-
-        return http.build();
-    }
-}
