--- examples/nostr-relay-example-application/src/main/java/org/tbk/nostr/relay/example/domain/event/EventEntityEvents.java
@@ -1,13 +1,0 @@
-package org.tbk.nostr.relay.example.domain.event;
-
-public final class EventEntityEvents {
-    private EventEntityEvents() {
-        throw new UnsupportedOperationException();
-    }
-
-    public record CreatedEvent(EventEntity.EventEntityId eventId) {
-    }
-
-    public record MarkDeletedEvent(EventEntity.EventEntityId eventId) {
-    }
-}
