+++ examples/nostr-relay-example-application/src/main/resources/application-development.yml
@@ -8,7 +8,7 @@ logging.level.org.springframework: INFO
 logging.level.org.springframework.web.servlet.mvc.method.annotation.RequestMappingHandlerMapping: TRACE
 logging.level.web: DEBUG
 logging.level.org.tbk.nostr: DEBUG
-logging.level.org.tbk.nostr.relay.example.domain.event: TRACE
+logging.level.org.tbk.nostr.example.relay.domain.event: TRACE
 
 spring.datasource.url: 'jdbc:sqlite:nostr_relay_example_application-dev.db?foreign_keys=true&journal_mode=WAL&synchronous=OFF&temp_store=MEMORY&page_size=32768&mmap_size=30000000000'
